utils: move Network operation helpers into apiserver
These two functions were only used in the apiserver. Move them
and their tests to the apiserver package.
